+++ src/bl64-gcp.bash
@@ -18,6 +18,8 @@ function bl64_gcp_run_gcloud() {
   bl64_check_parameters_none "$#" || return $?
   local debug=' '
   local config=' '
+  local project=' '
+  local impersonate_sa=' '
 
   bl64_check_module 'BL64_GCP_MODULE' ||
     return $?
@@ -29,7 +31,8 @@ function bl64_gcp_run_gcloud() {
   fi
 
   bl64_gcp_blank_gcloud
-
+  [[ -n "$BL64_GCP_CLI_PROJECT" ]] && project="--project=${BL64_GCP_CLI_PROJECT}"
+  [[ -n "$BL64_GCP_CLI_IMPERSONATE_SA" ]] && impersonate_sa="--impersonate-service-account=${BL64_GCP_CLI_IMPERSONATE_SA}"
   [[ "$BL64_GCP_CONFIGURATION_CREATED" == "$BL64_VAR_TRUE" ]] && config="--configuration $BL64_GCP_CONFIGURATION_NAME"
 
   bl64_dbg_lib_trace_start
@@ -37,6 +40,8 @@ function bl64_gcp_run_gcloud() {
   "$BL64_GCP_CMD_GCLOUD" \
     $debug \
     $config \
+    $project \
+    $impersonate_sa \
     "$@"
   bl64_dbg_lib_trace_stop
 }